refactor: :recycle: Réorganiser la structure des fichiers en déplaçant App.tsx dans le dossier components/App
