+++ src/components/App/App.tsx
@@ -1,11 +1,11 @@
 import { Outlet } from 'react-router-dom';
 import { useEffect } from 'react';
-import Footer from './components/Footer/Footer';
+import Footer from '../Footer/Footer';
 
 // import { CartProvider } from '../Cart/CartContext/CartContext';
-import { useUser } from './context/UserContext';
-import api from './api';
-import Header from './components/Header/Header';
+import { useUser } from '../../context/UserContext';
+import api from '../../api';
+import Header from '../Header/Header';
 
 function App() {
   const { setUser } = useUser();
